리뷰 좋아요 기능 개발

See merge request cloud_track/class_01/web_project3/team01/thisthat-be!8

+++ src/main/java/com/example/demo/web/result/ReviewResult.java
@@ -7,6 +7,6 @@ public class ReviewResult {
     private Long id;
     private String title;
     private String content;
-    private Integer star;
+    private Double star;
     private Integer heartCount;
 }
